% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Individual_level plots-tables.R
\name{compare.reaction.time}
\alias{compare.reaction.time}
\title{Compare participant reaction times}
\usage{
compare.reaction.time(dataframe, save_file_as = "rt_changes.csv")
}
\arguments{
\item{dataframe}{includes information about the participants}

\item{save_file_as}{name of the file we want to save as}
}
\value{
a saved csv file
}
\description{
generates a data frame for comparing reaction times across timepoints by individuals and words
}
\examples{
compare.reaction.time(final_processed_file,save_file_as='rt_changes.csv')
}
