% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Individual_level plots-tables.R
\name{list.by.timepoint}
\alias{list.by.timepoint}
\title{List participants by timepoint}
\usage{
list.by.timepoint(timepointcount, dataframe)
}
\arguments{
\item{timepointcount}{till what timepoint you want to list the participants by}

\item{dataframe}{contains information about participants}
}
\value{
saved csv file
}
\description{
generates a list of participants by timepoint
}
\examples{
list.by.timepoint(3,final_processed_file)
}
