% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Individual_level plots-tables.R
\name{plot.accuracy}
\alias{plot.accuracy}
\title{Plot accuracy charts}
\usage{
\method{plot}{accuracy}(dataframe, group_number, magicword_group, timepoint_count)
}
\arguments{
\item{dataframe}{includes information about the participants}

\item{group_number}{used to create the plots}

\item{magicword_group}{group number of the magic words}

\item{timepoint_count}{till what timepoint you want to list the participants by}
}
\value{
plots
}
\description{
this function finds the accuracy by word type (flags for class and time point) and also plots give the accuracy for each individual.
}
\examples{
plot.accuracy(final_processed_file,1,1,2)
}
