% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Individual_level plots-tables.R
\name{plot.class.time_point.percentages}
\alias{plot.class.time_point.percentages}
\title{Plot percentages by class and timepoint}
\usage{
\method{plot}{class.time_point.percentages}(dataframe, group_number, magicword_group)
}
\arguments{
\item{dataframe}{includes information about the participants}

\item{group_number}{is used to create the plots}

\item{magicword_group}{the group number of the magic words}
}
\value{
two plots
}
\description{
first plots the percentages of Responses by Word Type Across Timepoints for a specified class, and then plots one for all classes
}
\examples{
plot.class.time_point.percentages(final_processed_file,1,1)
}
