% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Individual_level plots-tables.R
\name{wide.or.long}
\alias{wide.or.long}
\title{Save in wide, Long, or both csv format}
\usage{
wide.or.long(format, dataframe)
}
\arguments{
\item{format}{of the csv file you want to save as (either 'wide','long', or 'both')}

\item{dataframe}{to be saved}
}
\value{
saved csv file
}
\description{
generates accuracy data for each individual and saves them in either or both wide and long format
}
\examples{
wide.or.long('both', final_processed_file)
}
